feat(scripts): Validate signature when downloading syncthing
The old Rakefile script from canton7 also used to do this, and it is
good practice to do so. The downloads are protected via HTTPS, but some
setups have less-than-trustworthy HTTPS proxies installed.

The new approach is slightly modernized, using SHA-256 instead of SHA-1.
The release key has also changed over the years, and needs to be
replaced.