fix(build): Build ProcessRunner as standalone exe
This is slightly ugly, as we can't currently build SyncTrayzor in a
single-file build due to CEF incompatibilities that I won't touch at the
moment. However, if ProcessRunner relies on the same .NET Core runtime
as SyncTrayzor, the auto-upgrade process can't work as the runtime files
are held locked during the upgrade.

Instead, use a different publish configuration for the (very
lightweight) ProcessRunner. However, this requires us to build it
manually.