feat(installer): Safely uninstall v1 installs on v2 upgrade. Prevent upgrades from 32-bit to 64-bit.
This change makes transitioning from v1 to v2 much easier. The new v2
installer will now detect if v1 is installed, and safely uninstall it.
It can optionally preserve the existing SyncTrayzor config (including
syncthing.exe), which is the default.

I can't be bothered to add the same logic for 32-bit installs (which is
more complicated anyway). Just prevent the installer from running if
someone is currently on the 32-bit v1 version. They will have to
uninstall and reinstall manually.