fix(ConflictResolution): Allow for non-existent base files in conflicts
Previously, SyncTrayzor required the "base file" for a conflict to
always exist, in addition to one or more conflicting files (e.g. a
conflict always requires 2+ files). This is no longer true for syncthing
v2: In v2, a deleted file can win a conflict, resulting in a single
conflicting file whose base file is deleted/nonexistent.

Properly handle this in code & GUI, and allow for such conflicts to be
resolved just like normal conflicts.