feat(installer): Allow for scripted uninstall
Scripted uninstalls, for example in choco, do not want user interaction.
The uninstaller however needs to know if it should remove all syncthing
configs or not. Previously, this was done via a message box which asked
the user.

This adds a new command line flag /DELETEALL that can be passed to Inno
Setup. If set, the message box is skipped and all syncthing
configuration data is deleted. If not set, the installer either asks the
user (if running interactively) or defaults to keep data (if running in
Inno's silent mode noninteractive).