feat: Warn user on startup if a mixed v1/v2 install is detected
We get frequent reports for users - in particular on the portable
version - who just mix v1 and v2 files. Since this causes incompatible
localization files to be merged, SyncTrayzor eventually crashes. This
upgrade procedure has never been supported by SyncTrayzor, but it likely
worked reasonably well enough in the past.

Warn if this is detected and instruct users to clean their install.

We also sometimes get reports of users on the installed version hitting
this problem. However, the v2 installer should remove v1 files, so it is
unclear why the installed version is sometimes affected by this as well.
A possibility is that it is caused by v1 files being locked by a stuck
v1 exe. In this case users also likely need to retry the installation.