fix(AboutView): Remove mention to the charity fundraising campaign
I had to remove this link from the repo unfortunately. It clashes with
the terms of service from GitHub regarding sponsorships and since it is
a fundraising managed by canton7 I do not have any control over where
the link points to. Since it no longer exists, remove the mention as
well.

The archived version of SyncTrayzor v1 still has the link, should anyone
wish to donate to the charity (if you read this, please check it out!).
I just cannot vouch for it, so unfortunately it no longer has a place in
v2.